refactor: 모듈 분리 및 비즈니스 로직 개선
- businessLogic.js: 할인 계산 및 포인트 로직 분리
- uiUpdates.js: DOM 조작 및 UI 렌더링 로직 분리
- main.basic.js: 이벤트 핸들링 및 앱 초기화 로직 정리
